+++ src/admin/admin.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+
+import { UserModule } from '../user/user.module';
+import { AdminModerationController } from './admin-moderation.controller';
+
+@Module({
+  imports: [UserModule],
+  controllers: [AdminModerationController],
+})
+export class AdminModule { }