+++ src/user/dto/update-role.dto.ts
@@ -1,0 +1,9 @@
+import { IsEnum } from 'class-validator';
+import { UserRole } from '../../common/enums/user-role.enum';
+import { ApiProperty } from '@nestjs/swagger';
+
+export class UpdateRoleDto {
+  @ApiProperty({ enum: UserRole })
+  @IsEnum(UserRole)
+  role: UserRole;
+}